Add support for decimal quantities in shopping list items
- Changed ShoppingItem quantity from int to double
- Added input formatter to limit decimal places to maximum 2
- Updated keyboard type to support decimal input
- Updated quantity parsing from int.tryParse to double.tryParse
- Updated quantity display formatting (shows whole numbers without decimals, decimals with up to 2 places)
- Updated CSV import to handle double quantities for backward compatibility
- Applied changes to both add and edit item dialogs

Users can now enter quantities like 0.5 liter, 0.25 kg, etc.

+++ lib/services/csv_service.dart
@@ -145,7 +145,7 @@ class CsvService {
           final name = fields[0].trim();
           if (name.isEmpty) continue;
 
-          final quantity = int.tryParse(fields[1].trim()) ?? 1;
+          final quantity = double.tryParse(fields[1].trim()) ?? 1.0;
           
           // Parse unit (if present) or default to piece
           ShoppingUnit unit = ShoppingUnit.piece;
